Add ban appeals cog
This cog listens for embeds to be sent by the forms webhook for user's submitting ban appeals. When one is detected it fetches extra info from the forms API and starts a thread, ready for mods to deliberate.

When a descision is made, it allows for the generation of a response to the ban appeal, using canned responses and custom input.

I have made a skeleton for what will eventually become and automatic email to appealers. This is currently waiting for PyDis to move to a different email supplier, so that we can integrate with the new API.